refactor(General): :recycle: Remove legacy address processes api
