--- mocks/data/SERVICE_PROCESSES.json
@@ -1,42 +1,0 @@
-{
-  "results": [
-    {
-      "recType": "PROCESS",
-      "identity": "process-payment-1",
-      "parent": "site-1",
-      "startTime": 1674048729000000,
-      "endTime": 0,
-      "name": "process payment 1",
-      "parentName": "site 1",
-      "imageName": "gcr.io/google-samples/microservices-demo/process-payment-1:v0.3.0",
-      "groupName": "payment",
-      "groupIdentity": "group-payment",
-      "hostName": "10.242.0.5",
-      "sourceHost": "172.17.63.163",
-      "processBinding": "bound",
-      "processRole": "external",
-      "addresses": ["payment:5000@payment:5000"]
-    },
-    {
-      "recType": "PROCESS",
-      "identity": "process-payment-3",
-      "parent": "site-1",
-      "startTime": 1674048729000000,
-      "endTime": 0,
-      "name": "process payment 3",
-      "parentName": "site 1",
-      "imageName": "gcr.io/google-samples/microservices-demo/process-payment-3:v0.3.0",
-      "groupName": "payment",
-      "groupIdentity": "group-payment",
-      "hostName": "10.242.0.5",
-      "sourceHost": "172.17.63.163",
-      "processBinding": "bound",
-      "processRole": "external",
-      "addresses": ["payment:5000@payment:5000"]
-    }
-  ],
-  "status": "",
-  "count": 2,
-  "timeRangeCount": 2,
-  "totalCount": 2
-}
